pyproject.toml
[tool.poetry]
name = "rfp-ai"
version = "0.1.0"
description = "RFP assistant to help with rfp"
authors = ["Jayesh Patil <jayeshvpatil@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = "^0.11.0"
uvicorn = "^0.27.1"
fastapi = "^0.110.0"
google-cloud-aiplatform = "^1.42.1"
langchain-exa = "0.0.1"
exa-py = "1.0.7"
python-dotenv = "1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
